services/ai-agent/requirements.txt
flask==3.0.3
gunicorn==22.0.0
litellm==1.44.22
kafka-python==2.0.2
prometheus-client==0.20.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
